// frontend/src/app/core/api/model/academic-year-dto.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AcademicYearDto { 
    readonly id?: number;
    readonly createdAt?: string;
    readonly updatedAt?: string;
    label: string;
    isActive?: boolean;
    startDate: string;
    endDate: string;
}