// frontend/src/app/core/api/model/exam-session-dto.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExamSessionDto { 
    readonly id?: number;
    readonly createdAt?: string;
    readonly updatedAt?: string;
    sessionType: ExamSessionDto.SessionTypeEnum;
    startDate: string;
    endDate: string;
    termId: number;
}
export namespace ExamSessionDto {
    export const SessionTypeEnum = {
        Main: 'MAIN',
        Retake: 'RETAKE'
    } as const;
    export type SessionTypeEnum = typeof SessionTypeEnum[keyof typeof SessionTypeEnum];
}